refactor: Move NED config parameter to the NED module
