+++ src/main/java/org/opentripplanner/graph_builder/GraphBuilderDataSources.java
@@ -17,13 +17,13 @@ import org.opentripplanner.datastore.OtpDataStore;
 import org.opentripplanner.datastore.api.CompositeDataSource;
 import org.opentripplanner.datastore.api.DataSource;
 import org.opentripplanner.datastore.api.FileType;
+import org.opentripplanner.graph_builder.module.ned.parameter.DemExtractConfig;
+import org.opentripplanner.graph_builder.module.ned.parameter.DemExtractConfigBuilder;
 import org.opentripplanner.graph_builder.module.osm.parameters.OsmExtractParameters;
 import org.opentripplanner.graph_builder.module.osm.parameters.OsmExtractParametersBuilder;
 import org.opentripplanner.standalone.config.BuildConfig;
 import org.opentripplanner.standalone.config.CommandLineParameters;
 import org.opentripplanner.standalone.config.api.OtpBaseDirectory;
-import org.opentripplanner.standalone.config.feed.DemExtractConfig;
-import org.opentripplanner.standalone.config.feed.DemExtractConfigBuilder;
 import org.opentripplanner.standalone.config.feed.GtfsFeedConfig;
 import org.opentripplanner.standalone.config.feed.GtfsFeedConfigBuilder;
 import org.opentripplanner.standalone.config.feed.NetexFeedConfig;
