+++ src/main/java/org/opentripplanner/graph_builder/module/ned/parameter/DemExtractConfig.java
@@ -1,8 +1,9 @@
-package org.opentripplanner.standalone.config.feed;
+package org.opentripplanner.graph_builder.module.ned.parameter;
 
 import java.net.URI;
 import java.util.Objects;
 import java.util.Optional;
+import org.opentripplanner.standalone.config.feed.DataSourceConfig;
 
 /**
  * Configuration of a DEM extract. Example:
