+++ src/main/java/org/opentripplanner/graph_builder/module/ned/parameter/DemExtractConfigBuilder.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.standalone.config.feed;
+package org.opentripplanner.graph_builder.module.ned.parameter;
 
 import java.net.URI;
