+++ src/main/java/org/opentripplanner/standalone/config/feed/DemConfig.java
@@ -2,6 +2,9 @@ package org.opentripplanner.standalone.config.feed;
 
 import static org.opentripplanner.standalone.config.framework.json.OtpVersion.NA;
 
+import org.opentripplanner.graph_builder.module.ned.parameter.DemExtractConfig;
+import org.opentripplanner.graph_builder.module.ned.parameter.DemExtractConfigBuilder;
+import org.opentripplanner.graph_builder.module.ned.parameter.DemExtractsConfig;
 import org.opentripplanner.standalone.config.framework.json.NodeAdapter;
 
 /**
